Create Callback Context
This is a class that is responsible for delegating writing to another class.
It does this through a map containing callbacks for different stages of the
output process. More stages might need to be added later.
